[COMMON] fix: service interface 제거
